cocktails/serializers.py
from rest_framework import serializers
from .models import Cocktail, Ingredient

class IngredientSerializer(serializers.ModelSerializer):
    class Meta:
        model = Ingredient
        fields = '__all__'

class CocktailSerializer(serializers.ModelSerializer):
    ingredients = IngredientSerializer(many=True, read_only=True)

    class Meta:
        model = Cocktail
        fields = '__all__'
